controllers/krate/metadata: Move `list_reverse_dependencies()` to dedicated module
